Moved Interface editor helper models to models package
