+++ plugins/org.eclipse.fordiac.ide.debug.ui/src/org/eclipse/fordiac/ide/debug/ui/view/FBDebugViewEditPartFactory.java
@@ -23,10 +23,10 @@ import org.eclipse.fordiac.ide.debug.ui.view.editparts.InnerValueEntity;
 import org.eclipse.fordiac.ide.debug.ui.view.editparts.InputEventValueEditPart;
 import org.eclipse.fordiac.ide.debug.ui.view.editparts.InterfaceValueEditPart;
 import org.eclipse.fordiac.ide.debug.ui.view.editparts.InterfaceValueEntity;
-import org.eclipse.fordiac.ide.fbtypeeditor.editparts.AbstractContainerElement;
 import org.eclipse.fordiac.ide.fbtypeeditor.editparts.FBInterfaceEditPartFactory;
 import org.eclipse.fordiac.ide.fbtypeeditor.editparts.FBTypeEditPart;
 import org.eclipse.fordiac.ide.fbtypeeditor.editparts.InterfaceContainerEditPart;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.AbstractContainerElement;
 import org.eclipse.fordiac.ide.model.libraryElement.FBType;
 import org.eclipse.fordiac.ide.model.libraryElement.With;
 import org.eclipse.gef.EditPart;
