+++ plugins/org.eclipse.fordiac.ide.fbtypeeditor/META-INF/MANIFEST.MF
@@ -6,6 +6,7 @@ Export-Package: org.eclipse.fordiac.ide.fbtypeeditor,
  org.eclipse.fordiac.ide.fbtypeeditor.editors,
  org.eclipse.fordiac.ide.fbtypeeditor.editparts,
  org.eclipse.fordiac.ide.fbtypeeditor.figures,
+ org.eclipse.fordiac.ide.fbtypeeditor.model,
  org.eclipse.fordiac.ide.fbtypeeditor.policies,
  org.eclipse.fordiac.ide.fbtypeeditor.properties
 Require-Bundle: org.eclipse.ui,
