+++ plugins/org.eclipse.fordiac.ide.fbtypeeditor/src/org/eclipse/fordiac/ide/fbtypeeditor/editparts/FBTypeEditPart.java
@@ -28,6 +28,15 @@ import org.eclipse.emf.common.notify.Notification;
 import org.eclipse.emf.common.notify.Notifier;
 import org.eclipse.emf.ecore.util.EContentAdapter;
 import org.eclipse.fordiac.ide.fbtypeeditor.figures.FBTypeFigure;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.AbstractContainerElement;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.EventInputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.EventOutputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.PlugContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.SocketContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.VarInOutInputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.VarInOutOutputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.VariableInputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.VariableOutputContainer;
 import org.eclipse.fordiac.ide.gef.editparts.AbstractConnectableEditPart;
 import org.eclipse.fordiac.ide.gef.listeners.DiagramFontChangeListener;
 import org.eclipse.fordiac.ide.model.libraryElement.AdapterType;
