+++ plugins/org.eclipse.fordiac.ide.fbtypeeditor/src/org/eclipse/fordiac/ide/fbtypeeditor/editparts/InterfaceContainerEditPart.java
@@ -22,6 +22,15 @@ import org.eclipse.draw2d.ToolbarLayout;
 import org.eclipse.draw2d.geometry.Dimension;
 import org.eclipse.emf.common.notify.Adapter;
 import org.eclipse.emf.common.notify.Notification;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.AbstractContainerElement;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.EventInputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.EventOutputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.PlugContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.SocketContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.VarInOutInputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.VarInOutOutputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.VariableInputContainer;
+import org.eclipse.fordiac.ide.fbtypeeditor.model.VariableOutputContainer;
 import org.eclipse.fordiac.ide.fbtypeeditor.policies.EventInputContainerLayoutEditPolicy;
 import org.eclipse.fordiac.ide.fbtypeeditor.policies.EventOutputContainerLayoutEditPolicy;
 import org.eclipse.fordiac.ide.fbtypeeditor.policies.InterfaceElementSelectionLayoutPolicy;
