+++ plugins/org.eclipse.fordiac.ide.fbtypeeditor/src/org/eclipse/fordiac/ide/fbtypeeditor/model/EventInputContainer.java
@@ -1,6 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2023 Profactor GmbH, fortiss GmbH,
- * 							Primetals Austria GmbH
+ * Copyright (c) 2011 - 2017 Profactor GmbH, fortiss GmbH
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -11,23 +10,22 @@
  * Contributors:
  *   Gerhard Ebenhofer, Alois Zoitl
  *     - initial API and implementation and/or initial documentation
- *   Alois Zoitl - copied and adapted from VariableOutputContainer
  *******************************************************************************/
-package org.eclipse.fordiac.ide.fbtypeeditor.editparts;
+package org.eclipse.fordiac.ide.fbtypeeditor.model;
 
 import java.util.List;
 
 import org.eclipse.fordiac.ide.model.libraryElement.FBType;
 import org.eclipse.fordiac.ide.model.libraryElement.IInterfaceElement;
 
-public class VarInOutOutputContainer extends AbstractContainerElement {
+public class EventInputContainer extends AbstractContainerElement {
 
-	public VarInOutOutputContainer(final FBType fbtype) {
+	public EventInputContainer(final FBType fbtype) {
 		super(fbtype);
 	}
 
 	@Override
 	public List<? extends IInterfaceElement> getChildren() {
-		return getFbType().getInterfaceList().getOutMappedInOutVars();
+		return getFbType().getInterfaceList().getEventInputs();
 	}
 }
