+++ plugins/org.eclipse.fordiac.ide.fbtypeeditor/src/org/eclipse/fordiac/ide/fbtypeeditor/model/VarInOutOutputContainer.java
@@ -1,5 +1,6 @@
 /*******************************************************************************
- * Copyright (c) 2011 - 2017 Profactor GmbH, fortiss GmbH
+ * Copyright (c) 2011, 2023 Profactor GmbH, fortiss GmbH,
+ * 							Primetals Austria GmbH
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,22 +11,23 @@
  * Contributors:
  *   Gerhard Ebenhofer, Alois Zoitl
  *     - initial API and implementation and/or initial documentation
+ *   Alois Zoitl - copied and adapted from VariableOutputContainer
  *******************************************************************************/
-package org.eclipse.fordiac.ide.fbtypeeditor.editparts;
+package org.eclipse.fordiac.ide.fbtypeeditor.model;
 
 import java.util.List;
 
 import org.eclipse.fordiac.ide.model.libraryElement.FBType;
 import org.eclipse.fordiac.ide.model.libraryElement.IInterfaceElement;
 
-public class VariableOutputContainer extends AbstractContainerElement {
+public class VarInOutOutputContainer extends AbstractContainerElement {
 
-	public VariableOutputContainer(final FBType fbtype) {
+	public VarInOutOutputContainer(final FBType fbtype) {
 		super(fbtype);
 	}
 
 	@Override
 	public List<? extends IInterfaceElement> getChildren() {
-		return getFbType().getInterfaceList().getOutputVars();
+		return getFbType().getInterfaceList().getOutMappedInOutVars();
 	}
 }
